fixed strlcat

+++ ft_strlcpy.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jarregui <jarregui@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jarregui <jarregui@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/22 15:28:20 by jarregui          #+#    #+#             */
-/*   Updated: 2022/11/02 18:57:38 by jarregui         ###   ########.fr       */
+/*   Updated: 2022/11/15 14:35:49 by jarregui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int size)
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	unsigned int	length;
-	unsigned int	i;
+	size_t	length;
+	size_t	i;
 
 	length = 0;
 	i = 0;
