+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jarregui <jarregui@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/28 09:29:29 by jarregui          #+#    #+#             */
-/*   Updated: 2021/07/14 22:53:23 by jarregui         ###   ########.fr       */
+/*   Updated: 2022/11/15 14:36:53 by jarregui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
